Separate apachethrift and http servers
