+++ crossdock/server/apachethrift/gauntlet.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package http
+package apachethrift
 
 import (
 	"errors"
